Rename `validate_csv.py` to `csv.py`
