+++ multinet/validation/csv.py
